moved home, not-found & its components to `(frontend)/(home)`
