+++ src/app/(frontend)/(home)/_components/Carousel.tsx
