+++ src/app/(frontend)/(home)/_components/Hero.tsx
