+++ src/app/(frontend)/(home)/_components/SponsorBubbles.tsx
