+++ src/app/(frontend)/(home)/_components/TextMarquee.tsx
