+++ src/app/(frontend)/(not-found)/not-found.tsx
@@ -8,10 +8,10 @@ export default function notFound() {
       <div className="flex space-x-6 text-white">
         {' '}
         {/* Links to other pages for navigation */}
-        <Link href="/" className="hover:underline">
+        <Link href="/public" className="hover:underline">
           Go to Home
         </Link>
-        <Link href="/" className="hover:underline">
+        <Link href="/public" className="hover:underline">
           About Us
         </Link>
         <Link href="/contact" className="hover:underline">
