--- src/app/my-route/route.ts
@@ -1,14 +1,0 @@
-import configPromise from '@payload-config'
-import { getPayload } from 'payload'
-
-export const GET = async () => {
-  const payload = await getPayload({
-    config: configPromise,
-  })
-
-  const data = await payload.find({
-    collection: 'users',
-  })
-
-  return Response.json(data)
-}
